deflate: rewrite the loops
Rewrite the loops so that gcc can vectorize them using saturated-sub.

The performance was improved by ~7% on my Linux box with an i7 CPU.

The original loop is legal to be vectorized;
gcc 4.7.* and 4.8.* somehow fail to catch this case.
There are still some room to squeeze from the vectorized code.

However, since these loops now account for about 1.5% of execution time,
it is not worthwhile to squeeze the performance via hand-writing assembly.

Change-Id: I13838555a2297bf12761371f6c699a98c09694eb
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>